Refactor content scripts.
