+++ src/ts/content/utils.ts
@@ -1,0 +1,35 @@
+import { PageContent } from "../common/types";
+import { buildSpecFrom } from "../parsers";
+import { OpenAPIObject } from "openapi3-ts";
+import debug from "../common/logging";
+
+const debugLog = debug("unmock:content:utils");
+
+let apiCheckResult: boolean;
+
+/**
+ * Check if can build a non-trivial OpenAPI spec from the page content. Cache the result to `apiCheckResult` to avoid
+ * building spec everytime this is called. This could mean missing some paths not loaded when this is first executed.
+ */
+export const checkIfCanParsePathsFromPage = async (): Promise<boolean> => {
+  if (typeof apiCheckResult !== "undefined") {
+    debugLog("Cached API check result", apiCheckResult);
+    return apiCheckResult;
+  }
+  try {
+    const pageContent = getPageContent();
+    const spec: OpenAPIObject = await buildSpecFrom(pageContent);
+    const isApi = Object.keys(spec.paths).length > 0;
+    apiCheckResult = isApi;
+  } catch (err) {
+    console.error("Failed building OpenAPI spec from page", err);
+  }
+  return !!apiCheckResult;
+};
+
+export const getPageContent = (): PageContent => {
+  const body = document.body;
+  const textContent = body.innerText || body.textContent;
+  const innerHtml = document.documentElement.innerHTML;
+  return { title: document.title, innerHtml, textContent };
+};
